Use quantile based computations for CZI range
Update pytools to include improved computations for the histogram
related parameters for multichannel images. The update improves
robustness to outliers and the background. And adds a qualtile based
range method which can be used instead of the MAD range for the range
of the pixel intensities to visualize in neuroglancer.